requirements.txt
Django~=5.1.6
drf-yasg~=1.21.9
djangorestframework~=3.15.2
django-environ~=0.12.0